feat: teach DateTimePartsArray to cast
I also:
1. Moved all the canonical code into canoncial.rs
2. Split the big test into smaller tests and put the smaller tests into the scrutinized files.
3. Used rstest for parameterized tests.
