+++ encodings/datetime-parts/src/lib.rs
@@ -2,6 +2,7 @@ pub use array::*;
 pub use compress::*;
 
 mod array;
+mod canonical;
 mod compress;
 mod compute;
 mod stats;
